Dropdown menu
Changed to a drop down menu.
lots of javascript was added to create the dropdown, first to enable
the dropdown menu, and then to call the different tabs.

The html was changed to become a dropdown, by creating a single li,
which has another ul nested inside of it. The ul has the links needed,
noted as a “#” followed by a, b, or c.

The css changed to make the text saying “dropdown” black when hovered
over by the mouse.